created project section
The project section is created with the dummy data, it was made responsive as well. primarily, there are some issues, but it can be fixed later. ***Also need to be mentioned that, the odd numbered items in the project section should be in a zig zag manner. Which means, the in one item, image should sit left and in the next one, it should sit right and so on.